requirements.txt
google-generativeai>=0.3.0
websocket-client>=1.6.0
mcstatus>=11.0.0
aiohttp>=3.9.0
requests>=2.31.0
